backend/requirements.txt
functions-framework==3.8.1
python-dotenv==1.0.0
requests==2.31.0
openai==2.30.0
boto3==1.34.0
google-auth==2.27.0
google-api-python-client==2.118.0
google-auth-oauthlib==1.2.0
